refactor: extract prefer-source-imports resolution
